First part of the refactor2
